order module completed

+++ The_Green_Country/src/main/java/com/masai/models/Planter.java
@@ -52,7 +52,7 @@ public class Planter {
 	
 	@NotNull(message = "Planter Cost should not be null")
 	@Min(value = 1, message = "Planter cost is must be greater than equal to 1")
-	private Integer PlanterCost;
+	private Double PlanterCost;
 	
 	@OneToMany(cascade = CascadeType.ALL)
 	private List<Seed> seeds = new ArrayList<>();
